Setup basic file
the file is now functioning on Vec2 rather than Vector2, the game will create a scene which will use the singleton method to allow for easy access of the UILayer, Background and Workspace objects.

I also changed the == operator for GameObjects, they will now first check if an object has been destroyed and if so then assume it has been deemed null.